writer/pipeline-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
supabase==2.9.1
python-json-logger==2.0.7
anthropic==0.36.1
openai==1.51.0
numpy==1.26.4
beautifulsoup4==4.12.3
lxml==5.3.0
google-cloud-language==2.14.0
nltk==3.9.1
